--- admin/person/delete_person.php
@@ -1,19 +1,0 @@
-<?php
-require_once "../../include/inc.php"; 
-
-if (isset($_POST["rep"])) {
-    if (isset($_GET["personID"]) && $_POST["rep"] == 1) {
-
-        $PersonRepo = new PersonRepository;
-        $person = $PersonRepo->getPersonById($_GET['personID']);
-        $name = $person->getName();
-        $PersonRepo->delete($person);
-        header("location:..\..\admin\list_movie.php?address=person&&del=$name");
-        exit;
-    } else {
-        header("location:..\..\admin\list_movie.php?address=person&&error=2");
-        exit;
-    }
-}
-
-require_once '..\..\template\admin\people\delete_person.phtml';
