+++ include/Controller/Admin/UserController.php
@@ -1,0 +1,27 @@
+<?php
+
+namespace Controller\Admin;
+
+use UserRepositoty;
+
+class UserController extends AdminController
+{
+    public function view()
+    {
+        $UserRepo = new \UserRepository();
+        $users = $UserRepo->getAll();
+
+        require_once '../template/admin/user.phtml';
+    }
+    public function add()
+    {
+    }
+
+    public function edit()
+    {
+    }
+
+    public function delete()
+    {
+    }
+}
