+++ include/Repository/PersonRepository.php
@@ -106,13 +106,14 @@ class PersonRepository
     public function update(Person $person)
     {
         $pdo = MyPDO::getInstance();
-        $query = 'UPDATE person SET name = :name, gender = :gender, birthday = :birthday, role = :role';
+        $query = 'UPDATE person SET name = :name, gender = :gender, birthday = :birthday, role = :role WHERE personID = :personID';
         $stmt = $pdo->prepare($query);
         $stmt->execute([
             ':name'=>$person->getName(),
             ':gender'=> $person->getGender(),
             ':birthday'=>$person->getBirthday(),
-            ':role'=>$person->getRole()
+            ':role'=>$person->getRole(),
+            ':personID' => $person->getPersonID()
         ]);
     }
