+++ template/admin/index.phtml
@@ -32,12 +32,10 @@
             <div class="main">
                 <a href="index.php">Trang chủ</a>
             </div>
-            <div class="main"><a href="list_movie.php?address=genre">Genre</a></div>
-            <div class="main"><a href="list_movie.php?address=person">Person</a></div>
-            <div class="main"><a href="list_movie.php?address=user">User</a></div>
-            <form method="post">
-                <button class="add" type="submit" name="add">Add film</button>
-            </form>
+            <div class="main"><a href="index.php?controller=admin.genre&action=view">Genre</a></div>
+            <div class="main"><a href="index.php?controller=admin.person&action=view">Person</a></div>
+            <div class="main"><a href="index.php?controller=admin.user&action=view">User</a></div>
+            <a href="index.php?controller=admin.film&action=add"><button class="add">Add film</button></a>
             <div class="clear"></div>
         </div>
         <div class="inform"> 
@@ -68,8 +66,8 @@
                 echo '<a href="'.$film->getImg().'">Img</a></td><td>';
                 echo '<a href="'.$film->getLink().'">Link</a></td><td>';
                 echo $film->getDescription().'</td><td>';
-                echo ('<a href="update_movie.php?id='.$film->getFilmID().'"><button class="button" type="submit">Update</button></a>');
-                echo ('<a href="delete_movie.php?id='.$film->getFilmID().'"><button class="button" type="submit">Delete</button></a>');
+                echo ('<a href="index.php?controller=admin.film&action=edit&id='.$film->getFilmID().'"><button class="button" type="submit">Update</button></a>');
+                echo ('<a href="index.php?controller=admin.film&action=delete&id='.$film->getFilmID().'"><button class="button" type="submit">Delete</button></a>');
                 echo ("\n</td></tr>\n");
             }
             echo '</table>';
