+++ template/admin/person.phtml
@@ -32,11 +32,11 @@
             <div class="main">
                 <a href="index.php">Trang chủ</a>
             </div>
-            <div class="main"><a href="list_movie.php?address=genre">Genre</a></div>
-            <div class="main"><a href="list_movie.php?address=person">Person</a></div>
-            <div class="main"><a href="list_movie.php?address=user">User</a></div>
+            <div class="main"><a href="index.php?controller=admin.genre&action=view">Genre</a></div>
+            <div class="main"><a href="index.php?controller=admin.person&action=view">Person</a></div>
+            <div class="main"><a href="index.php?controller=admin.user&action=view">User</a></div>
 
-            <a href="<?php echo '../admin/'.$_GET["address"].'/add_'.$_GET["address"].'.php'?>"><button class="add" type="submit">Add</button></a>
+            <a href="index.php?controller=admin.person&action=add"><button class="add" type="submit">Add</button></a>
             <div class="clear"></div>
         </div>
         <div>
@@ -52,17 +52,7 @@
         </div>
         <div class="films">
             <?php
-            if (isset($genres)) {
-                echo '<table><tr><th>GenreID</th><th>Name</th><th></th></tr>';
-                foreach ($genres as $genre) {
-                    echo '<tr><td>' . $genre->getGenreID() . '</td><td>';
-                    echo $genre->getName() . '</td><td>';
-                    echo ('<a href="..\admin\genre\update_genre.php?genreID='. $genre->getGenreID() . '"><button class="button" type="submit">Update</button></a>');
-                    echo ('<a href="..\admin\genre\delete_genre.php?genreID='. $genre->getGenreID() . '"><button class="button" type="submit">Delete</button></a>');
-                    echo '</td></tr>';
-                }
-                echo '</table>';
-            } elseif (isset($people)) {
+            
                 if (isset($_GET["del"]))
                 {
                     echo $_GET["del"]." deleted";
@@ -74,24 +64,11 @@
                     echo $person->getGender() . '</td><td>';
                     echo $person->getBirthday() . '</td><td>';
                     echo $person->getRole() . '</td><td>';
-                    echo ('<a href="..\admin\person\update_person.php?personID='. $person->getPersonID() . '"><button class="button" type="submit">Update</button></a>');
-                    echo ('<a href="..\admin\person\delete_person.php?personID='. $person->getPersonID() . '"><button class="button" type="submit">Delete</button></a>');
+                    echo ('<a href="index.php?controller=admin.person&action=edit&personID='. $person->getPersonID() . '"><button class="button" type="submit">Update</button></a>');
+                    echo ('<a href="index.php?controller=admin.person&action=delete&personID='. $person->getPersonID() . '"><button class="button" type="submit">Delete</button></a>');
                     echo '</td></tr>';
                 }
                 echo '</table>';
-            } elseif (isset($users)) {
-                echo '<table><tr><th>UserID</th><th>Name</th><th>User Name</th><th></th></tr>';
-                foreach ($users as $user) {
-                    echo '<tr><td>' . $user->getUserID() . '</td><td>';
-                    echo $user->getName() . '</td><td>';
-                    echo $user->getUserName() . '</td><td>';
-                    echo ('<a href="..\admin\user\update_user.php?userID=' . $user->getUserID(). '"><button class="button" type="submit">Update</button></a>');
-                    echo ('<a href="..\admin\user\delete_user.php?userID=' . $user->getUserID(). '"><button class="button" type="submit">Delete</button></a>');
-                    echo '</td></tr>';
-                }
-                echo '</table>';
-            }
-
             ?>
             <div class="clear"></div>
         </div>
